myblog/requirement.txt
python--3
django--2